% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{check_fitness}
\alias{check_fitness}
\title{Fitness function
Assess the fitness of each strategy through the use of a sequential game
between each focal strategy and a fixed number of random opponents}
\usage{
check_fitness(history, agents, pay, num_opponents = 10, rounds = 100,
  useC = FALSE)
}
\arguments{
\item{history}{A table of all possible prior moves of agents in sequence}

\item{agents}{A list of agents whose fitness will be assessed}

\item{num_opponents}{The number of random opponents to match the focal agent
against}

\item{rounds}{The number of rounds that will be played}
}
\value{
fitness The fitness that each agent accumlated
}
\description{
Fitness function
Assess the fitness of each strategy through the use of a sequential game
between each focal strategy and a fixed number of random opponents
}
