% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover function}
\usage{
crossover(agents, prob = 0.1)
}
\arguments{
\item{agents}{A list of agents on which crossing over will be performed}

\item{prob}{The probability that crossing over happens at a locus}
}
\value{
agents The list of agents after crossing over
}
\description{
Function that causes uniform crossing over among loci in a population of 
agents
}
