% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_ga_gui.R
\name{games_ga_gui}
\alias{games_ga_gui}
\title{Games genetic algorithm (GUI):}
\usage{
games_ga_gui()
}
\value{
A table (visible in a browser), the elements of which include: 
1. A table of the genomes of strategies and their frequencies in the 
population and 2. A plot showing the mean fitness calculated over all players 
in the population in each generation. The fitness of one player is the number 
of points accrued over all rounds and opponents in a generation.
}
\description{
This function launches a graphical user interface (GUI) within a browser for 
the games_ga function, which runs a genetic algorithm that identifies 
sequential strategies for maximising payoffs given any two by two symmetrical 
payoff matrix. Simulated players remember three rounds into the past. 
Outputs include a table of survivng adaptive strategies and a plot of mean
strategy fitness over generations of the genetic algorithm. No arguments are
required (or available) for this function -- it only launces the GUI.
}
