% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutation}
\alias{mutation}
\title{Mutation function
Function that causes alleles to mutate in agents}
\usage{
mutation(agents, prob = 0.01)
}
\arguments{
\item{agents}{A list of agents on which mutation will occur}

\item{prob}{The probability that a mutation will occur for any locus}
}
\value{
agents The list of agents after mutation
}
\description{
Mutation function
Function that causes alleles to mutate in agents
}
