% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament}
\alias{tournament}
\title{Tournament function
Function that simulates selection of next generation of agents according to
their fitness}
\usage{
tournament(agents, fitness)
}
\arguments{
\item{agents}{A list of agents to be assessed by fitness}

\item{fitness}{The fitness vector on which agents will be assessed}
}
\value{
agents A new list of agents selected according to fitness
}
\description{
Tournament function
Function that simulates selection of next generation of agents according to
their fitness
}
